requirements.txt
asgiref==3.3.4
Django==3.2.4
gunicorn==20.1.0
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
